Create basic web-scraper functionality.
This commit initializes the architecture for a web-scraper api. The
scraping is done asynchronously using the default Rails ActiveJob
enqueueing system (aka in-process memory). It is written to a static
file in lib/data that can be overwritten at any time. The intended
purpose of the `scrape` endpoint is to initiate a scrape for a client,
and return a status of the scrape (whether initialized or failure case).
The Extractors were meant to be multiplied in proportion to the needs
that arise, and/or grouped together if need-be.