Refactor scrape schema to use relations instead of embedded documents
Trying to embed links was a nice learning exercise, but I encountered
a use case for wanting to be able to extract all links across scrapes.

I also tweaked the extractors so that the link extractor has knowledge
of the URL being scraped. This means I can grab its parts to build
fully qualified links when relative links are encountered.